Remove superfluous code from strophe.caps.js.
The addFeature/removeFeature wrappers are redundant;
clients should simply call the disco functions directly.

The addIdentity() in init() is obnoxious as it cannot
be turned off or removed; the client should handle this.

Replace sendPres with a wrapper for $pres that takes attributes
and returns the stanza for further processing.

This allows the <c/> annotation to be added to a directed
presence as well.

(Renamed CAPS -> caps for consistency.)